+++ src/test/java/com/flab/readnshare/domain/favorite/controller/FavoriteControllerTest.java
@@ -3,17 +3,19 @@ package com.flab.readnshare.domain.favorite.controller;
 import com.flab.readnshare.domain.book.dto.BookDto;
 import com.flab.readnshare.domain.favorite.service.FavoriteBookService;
 import com.flab.readnshare.domain.member.domain.Member;
-import com.flab.readnshare.global.common.resolver.SignInMember;
+import com.flab.readnshare.global.common.advice.ApiExceptionAdvice;
+import com.flab.readnshare.global.common.exception.FavoriteException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mockito;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.context.annotation.Import;
 import org.springframework.core.MethodParameter;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
@@ -23,32 +25,39 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.*;
+import static org.mockito.BDDMockito.*;
+import static org.mockito.Mockito.verify;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @ExtendWith(MockitoExtension.class)
-@WebMvcTest(FavoriteController.class)
 class FavoriteControllerTest {
-    private MockMvc mockMvc;
-    private FavoriteBookService favoriteBookService;
+
+    static final String FAVORITES_API_ENDPOINT = "/api/v1/favorites";
+
+    @Mock
+    FavoriteBookService favoriteBookService;
+
+    @InjectMocks
+    FavoriteController favoriteController;
+
+    @InjectMocks
+    ApiExceptionAdvice apiExceptionAdvice;
+
+    MockMvc mockMvc;
 
     @BeforeEach
-    void setUp() {
-        // 필요한 서비스 모킹
-        favoriteBookService = mock(FavoriteBookService.class);
-        FavoriteController favoriteController = new FavoriteController(favoriteBookService);
-
-        // @SignInMember를 처리할 커스텀 인자 해결기 등록
+    void init() {
+        // @SignInMember 어노테이션을 처리할 커스텀 인자 해결기 등록
         HandlerMethodArgumentResolver signInMemberResolver = new HandlerMethodArgumentResolver() {
             @Override
             public boolean supportsParameter(MethodParameter parameter) {
-                return parameter.hasParameterAnnotation(SignInMember.class);
+                return parameter.hasParameterAnnotation(com.flab.readnshare.global.common.resolver.SignInMember.class);
             }
-
             @Override
             public Object resolveArgument(MethodParameter parameter,
                                           ModelAndViewContainer mavContainer,
@@ -60,55 +69,121 @@ class FavoriteControllerTest {
 
         mockMvc = MockMvcBuilders.standaloneSetup(favoriteController)
                 .setCustomArgumentResolvers(signInMemberResolver)
+                .setControllerAdvice(apiExceptionAdvice)
                 .build();
     }
 
-    @Test
-    @DisplayName("즐겨찾기 추가 API 테스트")
-    void addFavorite() throws Exception {
-        String isbn = "12345";
+    @Nested
+    @DisplayName("addFavorite 테스트")
+    class AddFavoriteTest {
+        @Test
+        @DisplayName("성공 - 즐겨찾기 추가")
+        void success() throws Exception {
+            String isbn = "12345";
 
-        mockMvc.perform(post("/api/v1/favorites")
-                        .param("isbn", isbn))
-                .andExpect(status().isCreated());
+            // 서비스에서는 예외 없이 정상적으로 처리한다고 가정
+            ResultActions resultActions = mockMvc.perform(
+                    post(FAVORITES_API_ENDPOINT)
+                            .param("isbn", isbn)
+            );
 
-        verify(favoriteBookService).addFavorite(argThat(member -> member.getId() == 1L), eq(isbn));
+            resultActions
+                    .andDo(print())
+                    .andExpect(status().isCreated());
+            verify(favoriteBookService).addFavorite(any(Member.class), eq(isbn));
+        }
+
+        @Test
+        @DisplayName("실패 - 즐겨찾기 추가 중 예외 발생")
+        void fail_addFavorite() throws Exception {
+            String isbn = "12345";
+            // 예: 이미 즐겨찾기에 등록된 경우 FavoriteAlreadyExist 예외 발생
+            willThrow(new FavoriteException.FavoriteAlreadyExist())
+                    .given(favoriteBookService).addFavorite(any(Member.class), eq(isbn));
+
+            ResultActions resultActions = mockMvc.perform(
+                    post(FAVORITES_API_ENDPOINT)
+                            .param("isbn", isbn)
+            );
+
+            resultActions
+                    .andDo(print())
+                    .andExpect(status().isBadRequest())
+                    // 예외 처리 Advice에서 전달하는 JSON 형식의 메시지 검증 (메시지는 상황에 따라 변경)
+                    .andExpect(jsonPath("$.message").value("이미 즐겨찾기에 등록되어 있는 책입니다."));
+        }
     }
 
-    @Test
-    @DisplayName("즐겨찾기 삭제 API 테스트")
-    void deleteFavorite() throws Exception {
-        String isbn = "12345";
+    @Nested
+    @DisplayName("deleteFavorite 테스트")
+    class DeleteFavoriteTest {
+        @Test
+        @DisplayName("성공 - 즐겨찾기 삭제")
+        void success() throws Exception {
+            String isbn = "12345";
+            willDoNothing().given(favoriteBookService).deleteFavorite(any(Member.class), eq(isbn));
 
-        mockMvc.perform(delete("/api/v1/favorites")
-                        .param("isbn", isbn))
-                .andExpect(status().isOk());
+            ResultActions resultActions = mockMvc.perform(
+                    delete(FAVORITES_API_ENDPOINT)
+                            .param("isbn", isbn)
+            );
 
-        verify(favoriteBookService).deleteFavorite(argThat(member -> member.getId() == 1L), eq(isbn));
+            resultActions
+                    .andDo(print())
+                    .andExpect(status().isOk());
+            verify(favoriteBookService).deleteFavorite(any(Member.class), eq(isbn));
+        }
+
+        @Test
+        @DisplayName("실패 - 즐겨찾기 삭제 중 예외 발생")
+        void fail_deleteFavorite() throws Exception {
+            String isbn = "12345";
+            willThrow(new FavoriteException.FavoriteNotFound())
+                    .given(favoriteBookService).deleteFavorite(any(Member.class), eq(isbn));
+
+            ResultActions resultActions = mockMvc.perform(
+                    delete(FAVORITES_API_ENDPOINT)
+                            .param("isbn", isbn)
+            );
+
+            resultActions
+                    .andDo(print())
+                    .andExpect(status().isNotFound())
+                    .andExpect(jsonPath("$.message").value("즐겨찾기에 등록된 책이 없습니다."));
+        }
     }
 
-    @Test
-    @DisplayName("즐겨찾기 조회 API 테스트")
-    void getFavoriteBooks() throws Exception {
-        BookDto dummyBook = BookDto.builder()
-                .id(10L)
-                .isbn("12345")
-                .title("Test Book")
-                .image("imageUrl")
-                .description("Test Description")
-                .author("Test Author")
-                .publisher("Test Publisher")
-                .link("http://test.com")
-                .build();
-        List<BookDto> dummyList = List.of(dummyBook);
+    @Nested
+    @DisplayName("getFavoriteBooks 테스트")
+    class GetFavoriteBooksTest {
+        @Test
+        @DisplayName("성공 - 즐겨찾기 목록 조회")
+        void success() throws Exception {
+            BookDto bookDto = BookDto.builder()
+                    .id(10L)
+                    .isbn("12345")
+                    .title("Test Book")
+                    .image("imageUrl")
+                    .description("Test Description")
+                    .author("Test Author")
+                    .publisher("Test Publisher")
+                    .link("http://test.com")
+                    .build();
+            List<BookDto> dummyList = List.of(bookDto);
 
-        when(favoriteBookService.getFavoriteBooks(argThat(member -> member.getId() == 1L)))
-                .thenReturn(dummyList);
+            given(favoriteBookService.getFavoriteBooks(any(Member.class)))
+                    .willReturn(dummyList);
 
-        mockMvc.perform(get("/api/v1/favorites"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].id").value(10L))
-                .andExpect(jsonPath("$[0].isbn").value("12345"))
-                .andExpect(jsonPath("$[0].title").value("Test Book"));
+            ResultActions resultActions = mockMvc.perform(
+                    get(FAVORITES_API_ENDPOINT)
+            );
+
+            resultActions
+                    .andDo(print())
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$[0].id").value(10L))
+                    .andExpect(jsonPath("$[0].isbn").value("12345"))
+                    .andExpect(jsonPath("$[0].title").value("Test Book"));
+        }
     }
 }
